backend/requirements.txt
# Django & DRF
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0
pgvector>=0.2,<0.3

# Celery
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# Google Cloud
google-cloud-aiplatform>=1.40,<2.0
google-auth>=2.25,<3.0
google-auth-oauthlib>=1.2,<2.0
google-api-python-client>=2.111,<3.0

# Web scraping
playwright>=1.40,<2.0
readability-lxml>=0.8,<0.9
beautifulsoup4>=4.12,<5.0
lxml>=5.1,<6.0

# Utilities
python-dotenv>=1.0,<2.0
cryptography>=41.0,<42.0
Pillow>=10.2,<11.0
httpx>=0.26,<0.27

# Development
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
black>=24.1,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
